python-api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics>=8.3.30
opencv-python==4.8.1.78
pillow>=11.0.0
numpy<2
torch>=2.6.0
torchvision>=0.21.0
gunicorn==21.2.0
huggingface_hub>=0.24.0
